Code cleanup. Removing unused code and fixing some issues.
