+++ api/oauth_github_callback.go
@@ -1,21 +1,27 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
-	"golang.org/x/oauth2"
-	"io/ioutil"
+	"io"
 	"net/http"
 )
 
 func githubGetPrimaryEmail(accessToken string) (string, error) {
 	client := &http.Client{}
 	req, err := http.NewRequest("GET", "https://api.github.com/user/emails", nil)
+	if err != nil {
+		return "", err
+	}
 	req.Header.Add("Authorization", "token "+accessToken)
 	resp, err := client.Do(req)
+	if err != nil {
+		return "", err
+	}
 	defer resp.Body.Close()
 
-	contents, err := ioutil.ReadAll(resp.Body)
+	contents, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return "", errorCannotReadResponse
 	}
@@ -47,7 +53,7 @@ func githubCallbackHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	token, err := githubConfig.Exchange(oauth2.NoContext, code)
+	token, err := githubConfig.Exchange(context.Background(), code)
 	if err != nil {
 		fmt.Fprintf(w, "Error: %s", err.Error())
 		return
@@ -61,6 +67,10 @@ func githubCallbackHandler(w http.ResponseWriter, r *http.Request) {
 
 	client := &http.Client{}
 	req, err := http.NewRequest("GET", "https://api.github.com/user", nil)
+	if err != nil {
+		fmt.Fprintf(w, "Error: %s", err.Error())
+		return
+	}
 	req.Header.Add("Authorization", "token "+token.AccessToken)
 	resp, err := client.Do(req)
 	if err != nil {
@@ -69,7 +79,7 @@ func githubCallbackHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	defer resp.Body.Close()
 
-	contents, err := ioutil.ReadAll(resp.Body)
+	contents, err := io.ReadAll(resp.Body)
 	if err != nil {
 		fmt.Fprintf(w, "Error: %s", errorCannotReadResponse.Error())
 		return
