+++ api/oauth_gitlab_callback.go
@@ -1,10 +1,10 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
-	"golang.org/x/oauth2"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 )
@@ -19,7 +19,7 @@ func gitlabCallbackHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	token, err := gitlabConfig.Exchange(oauth2.NoContext, code)
+	token, err := gitlabConfig.Exchange(context.Background(), code)
 	if err != nil {
 		fmt.Fprintf(w, "Error: %s", err.Error())
 		return
@@ -33,7 +33,7 @@ func gitlabCallbackHandler(w http.ResponseWriter, r *http.Request) {
 	logger.Infof("%v", resp.StatusCode)
 	defer resp.Body.Close()
 
-	contents, err := ioutil.ReadAll(resp.Body)
+	contents, err := io.ReadAll(resp.Body)
 	if err != nil {
 		fmt.Fprintf(w, "Error: %s", errorCannotReadResponse.Error())
 		return
