+++ api/utils_http.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"reflect"
 )
@@ -12,7 +12,7 @@ type response map[string]interface{}
 // TODO: Add tests in utils_http_test.go
 
 func bodyUnmarshal(r *http.Request, x interface{}) error {
-	b, err := ioutil.ReadAll(r.Body)
+	b, err := io.ReadAll(r.Body)
 	if err != nil {
 		logger.Errorf("cannot read POST body: %v\n", err)
 		return errorInternal
