+++ api/utils_misc.go
@@ -1,14 +1,9 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"os"
 )
-
-func concat(a bytes.Buffer, b bytes.Buffer) []byte {
-	return append(a.Bytes(), b.Bytes()...)
-}
 
 func exitIfError(err error) {
 	if err != nil {
