+++ api/utils_sql.go
@@ -1,7 +1,5 @@
 package main
 
-import ()
-
 // scanner is a database/sql abstraction interface that can be used with both
 // *sql.Row and *sql.Rows.
 type sqlScanner interface {
